requirements.txt
torch>=2.0.0
numpy>=1.21.0
Pillow>=9.0.0
faiss-cpu>=1.7.4
open-clip-torch>=2.20.0
anthropic>=0.8.0
python-dotenv>=1.0.0
tqdm>=4.65.0
psutil>=5.9.0
sqlite3>=3.35.0